backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-dotenv>=1.0.0
email-validator>=2.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pillow>=10.4.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.7
asyncpg>=0.29.0
requests>=2.28.0
cloudinary>=1.36.0